+++ packages/primitives/src/components/TextArea/index.ts
@@ -1,0 +1,2 @@
+export { default as TextArea } from './textarea';
+export * from './textarea.types';
