+++ packages/primitives/src/components/TextArea/textArea.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 
-import TextArea from '.';
+import { TextArea } from '.';
 
 export default {
   title: 'Primitives/TextArea',
