+++ packages/primitives/src/components/TextArea/textarea.tsx
@@ -1,23 +1,6 @@
-import { ChangeEvent, FocusEvent, useId } from 'react';
+import { useId } from 'react';
 
-// TODO : Auto Resizing TextArea
-export type Resize = 'both' | 'none' | 'horizontal' | 'vertical' | 'auto';
-
-export type TextAreaProps = {
-  id?: string;
-  name?: string;
-  value?: string;
-  defaultValue?: string;
-  placeholder?: string;
-  cols?: number;
-  rows?: number;
-  required?: boolean;
-  resize?: Resize;
-  isError?: boolean;
-  isSuccess?: boolean;
-  onChange?: (e: ChangeEvent<HTMLTextAreaElement>) => void;
-  onBlur?: (e: FocusEvent<HTMLTextAreaElement>) => void;
-};
+import { TextAreaProps } from './textarea.types';
 
 /**
  *
@@ -39,7 +22,8 @@ export type TextAreaProps = {
  * ```
  */
 
-const TextArea = ({ id, isError, ...restProps }: TextAreaProps) => {
+const TextArea = (props: TextAreaProps) => {
+  const { id, isError, ...restProps } = props;
   const randomId = useId();
   const textAreaId = id ?? randomId;
   return (
